Move all source code under src
